requirements.txt
langchain
huggingface_hub
sentence-transformers
tiktoken
streamlit
youtube-transcript-api
docx2txt
pypdf
unstructured
faiss-cpu